% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{blend_colors}
\alias{blend_colors}
\title{Blend two colors with mixbox}
\usage{
blend_colors(x, y, mixratio = 0.5)
}
\arguments{
\item{x}{Color 1}

\item{y}{Color 2}

\item{mixratio}{Mix ratio, can be any float but is typically between 0 and 1.
Defaults to .5.}
}
\value{
A blended RGB color
}
\description{
Uses mixbox to blend two colors together
}
