% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramps.R
\name{blend_palette}
\alias{blend_palette}
\title{Create a discrete color palette}
\usage{
blend_palette(n, colors, values = NULL)
}
\arguments{
\item{n}{Number of colors to return}

\item{colors}{Colors to use}

\item{values}{Optional numeric vector between 0 and 1 if values should not
be equally spaced}
}
\value{
A character vector of color hex strings
}
\description{
Create a palette of size n based on blending colors
}
