% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramps.R
\name{pigmentRamp}
\alias{pigmentRamp}
\title{Create pigment-blended color ramp}
\usage{
pigmentRamp(colors, values = NULL)
}
\arguments{
\item{colors}{Character vector of colors}

\item{values}{Optional, values between 0,1 if not equally spaced}
}
\value{
A function that takes a numeric value and returns a blended color
}
\description{
Uses the mixbox library to create a color ramp/gradient
}
