% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIGMENTS.R
\docType{data}
\name{PIGMENTS}
\alias{PIGMENTS}
\title{Mixbox pigments}
\format{
\subsection{\code{PIGMENTS}}{

A named list of RGB colors as hex strings
}
}
\source{
\url{https://github.com/scrtwpns/mixbox}
}
\usage{
PIGMENTS
}
\description{
Named list for the pigments provided by the mixbox developers.
}
\keyword{datasets}
